+++ joulescope_ui/devices/jsdrv/js110.py
@@ -436,8 +436,9 @@ class Js110(Device):
                 'units': units,
                 'data': value['data'],
                 'origin_sample_id': value['sample_id'],
-                'orig_sample_freq': value['sample_rate'],
-                'orig_decimate_factor': value['decimate_factor'],
+                'origin_sample_freq': value['sample_rate'],
+                'origin_decimate_factor': value['decimate_factor'],
+                'time_map': value['time_map'],
             }
             fwd['data'] = value['data']
             self.pubsub.publish(utopic, fwd)