+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Bumped pyjls from 0.9.5 to 0.10.0.
 * Improved error handling when opening JLS file that does not exist.
 * Improved error reporter to show UI RAM utilization.
+* Added developer TimeSync widget to plot instrument synchronization.
 
 
 ## 1.1.10